refactor(paperlab)!: the public API is 83 names, not 214
Every exported name is a promise kept for years, and this library was exporting
its own internals. Shader builders, texture painters, tessellation constants,
the cloth integrator, the state machine class, and 36 individual behavior,
deformer and layout functions the registries already reach — none of it is API.
Shipping it means a refactor of a private helper becomes somebody's breaking
change.

What survives is what a caller genuinely cannot do without: the four
components, the schema and its types, the registries and their three
`register*` hooks, presets and the `.paper` format, the export helpers,
lighting-as-data, interaction states, accessibility.

Behaviors, deformers and layouts are reached through their registries.
`getBehavior('peel')` returns exactly what the `peel` export did. Nothing left
the library; it left the front door.

Three things that look internal stay, each with its reasoning written at the
export site — the parity harness (the only gate on the invariant the
contribution ladder rests on), the tessellation arithmetic (`registerDeformer`
is public, so a third-party deformer must answer the segment question the same
way the built-in seven do), and `wrapLines` (a caller measuring type before
layout has to get the painter's answer).

`PaperField.tsx` stops re-exporting the field internals it was forwarding by
habit; a re-export is how something becomes public by accident. Two test files
now import those from their own modules, which is where they always lived.

Timing is the whole argument. Nobody has built on the old surface, so this
costs no one anything today; the same change after launch is a migration
charged to real users for no benefit to them.

Verified: lint, knip, typecheck, 707 tests, build, publint, attw, and all three
browser gates — parity 37/37, drive, share 10/10.
